final submission
Unfortunately was not able to complete the app. I could not get the remove item from array list to work and wasn't sure how to search for items based on a name. I didn't manage to get the items shown based on stock level, but I don't think it was displayed as it was requested.